Fixed sonarqube issue
